proto: Add version field to STG proto format
In addition to the version field, the proto reader will maintain a list
of the supported proto versions. The reader will exit in case a proto
file with unsupported version is encountered.

The version field is mandatory for the format now. Since the field will
be missing from STG files which already have been written by the
writer, proto dictates that the version field will be assumed to have
the default value of 0. Hence, this maintains compatibility with
existing STG files, which will be considered version 0.

STG proto message fields are renumbered to keep the version field at
the top of the text format. Renumbering of fields does not break
compatibility with the text format since it is agnostic to the field
numbers.

PiperOrigin-RevId: 518532163
Change-Id: I4a0d48c43c0dc4adfb7d4d17d44290eaf820b7f5