type_roots: Emit `Interface` node from the proto writer
Upgrade the proto writer to use the `Interface` node while emitting the
proto STG format in place of the `Symbols` node.

This change to the written format is not compatible with version 0 of
the format. Hence, the proto writer will mark exported STG files with
version 1 from this change onwards.

PiperOrigin-RevId: 527232913
Change-Id: I4815a21e0ded29fdece618abf47d04d663fab480