Abigail reader: update is-declaration-only handling comment
Under rare circumstances, libabigail, or perhaps only older versions of it, can
emit non-empty struct types with the `is-declaration-only` flag set. The flag in
turn comes from DWARF's `DW_AT_declaration` which cannot be interpreted in
isolation.

We also cannot rely on just the absence of children to detect forward declations
as this does not distinguish `struct foo;` from `struct foo {};`.

In any case, this comment is now clearer and shorter.

PiperOrigin-RevId: 528750564
Change-Id: Ie30b0229db7c6779619fee7f491d965be854d4f9