DWARF processor: add C++ method support
* Add support for simple (non-virtual and non-static) C++ methods.
* Refactor data fetching out of `ProcessFunction` to reuse it in
  `ProcessMethod`.

At least with Clang v14 and DWARFv5 method actually has two DW_TAG_subprogram:
one inside DW_TAG_structure_type with “member” field, another is placed just
outside the structure with location information and DW_AT_specification. In
this change the first one is the one that used to construct `Method` node. In
our cases it doesn't have an address and is not passed to `result_.symbols`.

The second one is caught by `ProcessFunction`, it has an address and it is
passed to `result_.symbols`. But it doesn't have the name of structure attached
to the name of function (scope).  This will be fixed in following changes.

Both instances of DW_TAG_subprogram will produce two equal, but not connected
to each other `Function` nodes, which will be unified on later processing
stages. These nodes are artificial, not associated with `Dwarf_Off` and not
included in `std::unordered_map<Dwarf_Off, Id> id_map_`. That means that they
couldn't (and shouldn't) be referenced outside of containing `Method` or
`Types::Symbol`.

PiperOrigin-RevId: 537875694
Change-Id: Id361e8d7ce55df59921f3e293020230a5ad5e467