Add simple support for GNU_IFUNC (indirect) ELF function types
GNU_IFUNC function types are defined with
__attribute__((ifunc("resolve_func"))) and point to a resolver function to
determine the actual implementation at runtime. STG does currently not support
those types, hence add it to the model.

There is a caveat though that indirect functions do not come with a definition
and attached type information. Instead the resolved function has it. Similarly
the return type of the resolver function describes the type. Yet, those could
be implementation details hidden from the public interface, e.g. like this

 |  static void my_func() {}
 |
 |  static void (*resolve_func(void))(void) {
 |    return my_func;
 |  }
 |
 |  void func() __attribute__((ifunc("resolve_func")));

This change adds simple support to treat GNU_IFUNC like a function declaration
in order to at least track additions and removals. A follow-up change needs to
be made in order to model the type information of such functions.

PiperOrigin-RevId: 539022390
Change-Id: I0997699a4c3888d5b0335a6c4d43a27e8a3d81db