DWARF processing: allow missing linkage name
Set `stg::Method::mangled_name` to empty an string when we don't have a linkage
name in DWARF. This is the case when the method is constexpr and is always
inlined. Further, deleted functions do not have linkage names, but show up in
DWARF.

This is a workaround solution to unblock C++ ABI experiments. Proper support
for linkage names will be implemented later.

PiperOrigin-RevId: 539632733
Change-Id: I6e246226e87ccc51ce7792bb9874f9ab85a88212