unification: determine direction of substitutions
This is one of series of changes simplifying STG type resolution by eliminating
explicit tracking of forward-declared types.

This change makes `Unify` determine which of the nodes should be preserved in
the case of a substitution. This is currently the responsibility of
`UnificationCache` which needs to know which nodes corresponds to incomplete
types.

`Unify`'s function operator on nodes now returns two distinct values that mean
"can be unified", distinguishing whether the first or second type should be
preserved. As before, substitutions are added that preserve whichever type is
fully defined, preferring the second type if both are.

PiperOrigin-RevId: 539652024
Change-Id: I2ff7d67f9f87eceb9ff12f8473388a66d6f36282