type resolution: rename Unify to Unifier and add Unify wrapper function
Each recursive attempt at unification can be thought of as starting a
transaction. When the outermost unification succeeds all the pending
subsitutions can be committed, otherwise they must be discarded.

The new `Unify` function takes responsibility for committing successful
unification substitutions to the unification cache and is the new entry point
for unification.

This makes things simpler for the caller.

PiperOrigin-RevId: 539664071
Change-Id: I78d615d9c329f40a5f69cd52807faff4eaf32876