DWARF processor: add scoped name for functions
Functions are sometimes defined twice: the first inside the class, where it
serves as declaration of method or static field and the second outside of the
class where there is definition. The second refers to the first through
DW_AT_specification.

This change propagetes scoped name from entry with name to entry with
specification. First part is `GetNameWithContext` which extracts name and/or
specification for entry. If entry has name directly (not through
specification), it is wrapped into the current scope and stored in the
`scoped_names_` collection. `GetNameWithContext` is called for every subroutine
entry in DWARF.

Second part is using `NameWithContext` for assigning scoped name to the symbol.
It is done in `GetScopedNameForSymbol`. Either we already have scoped name,
which is immediately returned, or we have specification, which is registered as
"unresolved" in `unresolved_symbol_specifications_`.

After DWARF processing pass, all unresolved pairs "symbol-specification" are
resolved through the `scoped_names_` in `ResolveSymbolSpecifications`. Using
sort and linear pass is more efficient than resolving through
`std::unordered_map`, because "keys" (`Dwarf_Off`) are already almost sorted,
because processor goes through DWARF entries sequentially. `std::sort` is
needed to guarantee this invariant.

After this changes scope processing takes less than 100ms from 14.7s of a
kernel processing.

PiperOrigin-RevId: 540886421
Change-Id: Ie834c37c43f0ae4974ee827aa32b09ab233c1ce5