Abigail reader: report missing XML type definitions
Occasionally, due to split XML elements, some subelements do not get picked up
and this can result in missing member type definitions and a dangling edge in
the graph.

If the dangling edge is reachable, this will cause an `undefined node` error in
later processing. The new warning gives context for such failures.

PiperOrigin-RevId: 542223291
Change-Id: I13e199e2dac1f909d21194b92e9b5700b746a4b8