type resolution: rework ResolveTypes interface and callers
This change is a prerequisite to reworking type resolution and related things
within the ELF reader.

`ResolveTypes` is no longer responsible for creating the `Unification` object
or for updating nodes afterwards. `std::reference_wrapper` is also eliminated
from the type resolution interface.

In the ELF reader, the class `Typing` is merged into `Reader` and a single root
node ID is passed to `ResolveTypes`.

PiperOrigin-RevId: 543356061
Change-Id: I9a0ed53ba2339ad00e76343afc6ce98d30e1eba0