ELF reader: replace equality testing with unification
The ELF reader may find multiple types for one symbol. The types had
better be the same and this was enforced with an equality test. However, that
has some issues.

Unification is preferable to equality testing, as types may only differ due to
the presence of a forward declaration.

Using the equality and unification algorithms simultaneously is also a bad idea,
as substitutions by the latter are invisible to the former, until the graph is
rewritten.

PiperOrigin-RevId: 543447725
Change-Id: I065fb884ea48a119bbe601a362adca5be5527964