ELF reader: refactor root node building
This change renames the method `GetTypesFromDwarf` to `BuildRoot` and moves in
these bits of processing:

* symbol node creation
* root node creation

The method now owns the `Unification` object for its lifetime and this clearly
brackets the processing during which unification is active. It is called
exactly once by `Read`.

For consistency and to reduce argument passing, the vector of public functions
and variables is also made a member variable.

PiperOrigin-RevId: 544626193
Change-Id: I0ae321befe607313d776bc5c90c0ed5b2c1648f8