DWARF processor: treat DW_TAG_unspecified_type as Special::Kind::NULLPTR
DWARF can contain entries, that refer to the type, unknown in compilation type.

There is a special case that is easy to encounter: `decltype(nullptr)`. If this
is the case then `DW_TAG_unspecified_type` is interpreted as `Special` node
with kind `NULLPTR`. In other cases code will fail with "unsupported" error.

Add test with `decltype(nullptr)`.

PiperOrigin-RevId: 551217897
Change-Id: I49b5409d131fd75a9a48cf7f23725d43c8fab7ae