make Methods be exclusively virtual methods
These are the only kinds of methods that are tied to a class (via a
vtable). The others have independent existences as symbols.

Model changes:

`Method` node

* `kind` is removed
* `vtable_offset` is now non-optional

Format changes:

`method` node

* `kind` is removed along with all reader / writer support
* `vtable_offset` is now non-optional, so an absent field implies a 0
  offset
* "stable" IDs will differ

Note that the format version has not been bumped as there are no known
C++ STG ABIs in the wild.

PiperOrigin-RevId: 563736222
Change-Id: I37bb5a4e67f3d417492f0273ff5e6b4dff506bb4