Abigail reader: fix bad DWARF-ELF links caused by size zero symbol confusion cont.
libabigail used to be confused by these sorts of symbols, resulting in
declarations pointing at the wrong ELF symbols:

```
573623: ffffffc0122383c0   256 OBJECT  GLOBAL DEFAULT   33 vm_node_stat
573960: ffffffc0122383c0     0 OBJECT  GLOBAL DEFAULT   33 vm_numa_stat
```

The reader now also checks for variables where the ELF symbol ID and
mangled name are the same but the name is different which has been
seen in files in libabigail's own test suite.

In any cases, fixes are only applied to declarations that would
otherwise cause the Abigail reader to fail.

PiperOrigin-RevId: 564399127
Change-Id: Ib76f91bbf9a265a931085f8eef7814b2a3f0561c