comparison: assert matching pair invariants
When comparing two sequences of items, we generate a list of pairs of
optional values where at least one of the optional values is present.
Both values are present if the item is present in both sequences. Only
one value is present when there was no corresponding item in the other
sequence. It is impossible for both sides of the pair to be absent.

This is made a little clearer in the code now.

PiperOrigin-RevId: 565384098
Change-Id: I8c814e9ca4bec09af11b459d824f8d082dcc1bd3