Abigail reader: simplify check for duplicate symbol-type link
The code no longer allows duplicate links even if the relevant types
have the same IDs. Better deduplication logic has been added to the
Abigail reader making this no longer necessary (to the extent that it
was useful at all).

ABI XML models ELF symbol / DWARF declaration links as a
`symbol-type-id` attribute on a `var-decl` or `function-decl` element.
In older versions of `abidw` duplicate declarations (typically of
member variables and functions as part of a duplicate type
declaration) were often seen.

The code attempted to deal with such duplicates by recording which
type node ID was associated with each ELF symbol and only failing if
two distinct node IDs were recorded. However, this never worked for
the synthetic types generated for functions and also relied on node
identity which is not meaningful in general.

PiperOrigin-RevId: 566619013
Change-Id: I79a684272a3104667c93c2bd1814b6dc7f2a01e3