BTF reader: check alignment of structures more carefully
The entire BTF data could be misaligned, resulting in undefined
behaviour reading the BTF header.

The start of type information is provided by a field in the BTF
header. This could result in arbitrary alignment. The pointer, rather
than the offset, is now checked.

Also use `alignof` of structures instead of `sizeof(uint32_t)`.

PiperOrigin-RevId: 577812284
Change-Id: I3ff33a63ae8ec971a30a3db504bcbd454016e127